Partial commit of reduction changes

+++ include/dynd/func/reduction.hpp
@@ -57,9 +57,10 @@ namespace nd {
     callable reduction(const callable &elwise_reduction,
                        const ndt::type &lifted_arr_type,
                        const callable &dst_initialization, bool keepdims,
-                       intptr_t reduction_ndim, const std::vector<int> &axes,
-                       bool associative, bool commutative,
-                       bool right_associative, const array &reduction_identity);
+                       intptr_t reduction_ndim,
+                       const std::vector<intptr_t> &axes, bool associative,
+                       bool commutative, bool right_associative,
+                       const array &reduction_identity);
 
   } // namespace dynd::nd::functional
 } // namespace dynd::nd
